requirements.txt
asgiref==3.9.1
certifi==2025.8.3
distlib==0.4.0
dj-database-url==3.0.1
Django==5.2.5
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
filelock==3.19.1
gunicorn==23.0.0
mysqlclient==2.2.7
packaging==25.0
pillow==11.3.0
pipenv==2025.0.4
platformdirs==4.4.0
psycopg2-binary==2.9.11
Pygments==2.19.2
PyJWT==2.10.1
python-dotenv==1.1.1
setuptools==80.9.0
sqlparse==0.5.3
tzdata==2025.2
virtualenv==20.34.0
whitenoise==6.11.0
